Adding copy log

+++ Model/BackupLog.cs
@@ -9,11 +9,11 @@ namespace EasySave.Model
         public string Name { get; private set; }
         public string FileSource { get; private set; }
         public string FileTarget { get; private set; }
-        public int FileSize { get; private set; }
-        public double FileTransferTime { get; private set; }
+        public long FileSize { get; private set; }
+        public long FileTransferTime { get; private set; }
         public string time { get; private set; }
 
-        public BackupLog(string name, string sourceFile, string targetFile, int fileSize, double transfertTime, string time)
+        public BackupLog(string name, string sourceFile, string targetFile, long fileSize, long transfertTime, string time)
         {
             this.Name = name;
             this.FileSource = sourceFile;
